arm: msm: build: add option to build motorola dtbs
Introduce a new configuration MMI_DEVICE_DTBS that will allow the
target to only build Motorola devices. If this option is configured
we will omit the QC architecture defaults for msm8916. This also will
enable adding 1kb of padding to the DTB to allow for insertion of
variables into the /chosen node at runtime in the bootloader.

Support compiling off the 899x and 89x6 family of chips.

Change-Id: I038a6d77bd4b827c9881f59c9c61ac8c5c62d766
[jasonh: Adjusted for msm8916]
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>