esdfs: emulated SD card storage for Android
This commit implements the functionality of Android's FUSE-backed
"sdcard" file system using the wrapfs stackable file system approach.
The lower half overrides the caller's credentials to behave as the
Android user space "sdcard" service would on the underlying file
system.  The upper half implements a case-insensitive, case-preserving
file system with hard-coded permissions to emulate the behavior of the
vfat driver as much as possible.  To improve lookup performance, it
relies on the underlying file system to support the LOOKUP_NOCASE
intent.

Change-Id: Ifff10e0861dcc3fe491e96063a4ad67b62c947cd
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>