power: max17042 : Configure SOC & VLT thresholds for POR
The interrupt thresholds for SOC are programmed such that we get an
interrupt for every 'off' percentage change in the soc. When 0% is
reported by Fuel gauge at power-up, the lower threshold becomes -1
which is 0XFF. As a result of this, the fuel gauge interrupt line is
in pending state and blocks handling of other interrupts.

The interrupt thresholds for Voltage are programmed such that we get an
interrupt if the voltage drops below cut-off threshold. When powered up
at voltages below the cut-off threshold, fuel gauge interrupt line is in
pending state and blocks handling of other interrupts.

Change-Id: I135d77e8a5f9e8cabd46431d536c05c3aa650fd1
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.mot.com/676689
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>