misc: drv2605: Fix panic when drv2605 is disabled
If drv2605 is disabled in the device tree config module_init
is still called and /dev/drv260x is created. In this case
if the API will try to start a haptic playback the driver may
panic because the regulator and I2c client is not initialized.

Signed-off-by: Vladimir Karpovich <vkarpovich@motorola.com>

Change-Id: I0a05b0de372b093c2e17a86d785f8e3a1180c2f0
Reviewed-on: http://gerrit.mot.com/724003
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>