drivers:lmk: Fix null pointer issue
On some race, the tsk that lmk is using may be deleted from the RB tree
by other thread, and rb_next would return a NULL if we use this tsk to
get next. For this case, we need to skip this round of shrink and wait
for the next turn. Otherwise, tsk would trigger NULL pointer panic.

Change-Id: If6688d94a0b2992556cec0cf1847b58c69a38d8d
Signed-off-by: Hong-Mei Li <a21834@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>