lowmemorykiller: maintain LMK rbtree with signal->adj_node
Currently, we maintain LMK rbtree with task->adj_node. However, when
handling oom_score_adj change case, we may del/add a non-leader task
to the RB tree, which is not as expected.

This patch we maintain the LMK rbtree with task->signal->adj_node.
Since signal_struct is shared between main task and threads, we can
avoid non-leader thread adding to tree.

Change-Id: I3ba9e740e03ab04c25497a1cc2c870f051bd5b07
Signed-off-by: Hong-Mei Li <a21834@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>