config: msm: Migrate debug configs into ext_config
Synchronize the differences between msm-perf and msm defconfigs so that
when using the msm-perf configuration the ext_config/debug-msm.config
will overlay the debug features to turn it into the same as msm_ config.
This allows a single place to update common features and any debug
configurations are to be placed in the corresponding debug configuration
inside the ext_config folder. This commit brings us up to par with:

 49b07be Revert "kgsl: leave global domain attached"

New QCOM drops which update the msm_defconfig would need to be reflected
in later commits.

Change-Id: I484d50433c5bbc1ec0e94566eab8488d93fMontanadf
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>