arm/dt: 8996: Enable L18 as always on
Enable the L18 regulator and set initially for always-on. This powers
the circuit which OR's the vibrator enable from the AP and the sensor
hub. Eventually, this will need to be controlled in the display driver
as it also supplies VCI. It'll will need to be turned off for ESD
recovery.

Change-Id: I21cde4cb91afdbbabd9290235ee43fe7f8327e46
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>