power: qpnp-smbcharger-mmi: Prevent 0% Flag
There is a flag that can get set from an IRQ and this
will hold 0% shown to the User indefinitely when going
through Dead Battery recovery with a Turbo charger.

Prevent this by Polling the Raw Capacity for a value
larger then 0% so that it can be shown to the user.

Change-Id: I63368eac2b4b04d05b8978151904bd131121bac7
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/787957
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>