arm/dt: 8996: Princeton: Bump muc deKinzie to 3s
Increase from 500ms to 3s the deKinzie on princeton wingboards when
inserting a mod. The connector is difficult to set quickly and ensure
all the pins are mated before the communication is attempted to start.
We'll keep the 500ms deKinzie on closed up phones where the connection
is quicker/reliable.

Change-Id: Icac8953ad9eb73a727702c0e2891306c0b675c53
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.mot.com/799213
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>