power: qpnp-smbcharger-mmi: Add BSW Support
This adds the ability to detect and configure a USBC
Charger for Current Limit steps.  In combination with
multiple steps this also uses a bsw to connect the
battery and the charger.  That bsw can be set with
variable voltage trip points as well as the FG can be
used for a trip contdition as well.

Change-Id: I098fcd1f507630b43fcba153fb22a08e65e7cec8
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/825881
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>