arm: dts: Add SPI MOSI/MISO to muc gpios, add ack pin config
The SPI MOSI and MISO lines are used to ack messages to and from
the MuC. As a result, a new pin config is required so set MOSI
and MISO as GPIOs. In order to be able to read or assert an ACK,
the lines must also be passed as GPIOs.

Change-Id: Ia66189497b68aca4f6c133bc28c108d0b3ab79ce
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/828327
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
SME-Granted: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jason Hrycay <jason.hrycay@motorola.com>
Tested-by: Jason Hrycay <jason.hrycay@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>