power: qpnp-fg: Optimize Profile Loading
At Boot when the FG is booting there is a period of
time where it is unsure if it needs to reload the Battery
Profile.  During this time a Default of 50 is reported
for SOC.  That can lead to some confusion since due to
another rework this can take some time.

Redesign the Default SOC to be read straight from the
Raw SOC register.  This value, while not guaranteed to
be accurate should be closer to the final value then
the 50.  Also move the Charge Disable 2 second wait
after the check for needed Profile reload.  This will
allow most power-ups to move quicker to the Loaded State.

Change-Id: I6172007118185156731aaaa96848a127bfc9e6cd
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/838729
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Ravikumar Vembu <raviv@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>