video: Add panel_notifier framework
This is designed to allow other drivers to know about panel events. It
also enables drivers to inject synchronous operations into the panel
state flow, by blocking the panel state changes at various points.

An example of this would be if a driver would like to know when a panel
is on and do an operation before the video stream to that panel begins.

Change-Id: I319c0a3fb933f47bb3d9a5afedae268294fe144e
Signed-off-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-on: http://gerrit.mot.com/853489
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Joseph Swantek <jswantek@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>