slimport: anx7816: log/write 7751 BPC only if different
For debugging purposes, it would be nice to know when devices are using
the workaround. Compare the SP_TX_VID_CTRL2_REG IN_BPC (mask 0x70) of
the 7816 and 7751, and only log a message and write the register if the
values are different.

Change-Id: I9e475bcc93a94221522813e8a51190a3f4fe6817
Signed-off-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-on: http://gerrit.mot.com/865283
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Joseph Swantek <jswantek@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>