slimport: anx7805: add sw timeout for hdcp auth failed
Analogix has indicated that there is a known issue with HDCP
authentication sometimes "not happening". Newer chips such as the 7816
that we use on the 8996 platform have internal timeout mechanisms that
they use to interrupt the MSM and restart the HDCP auth process. The
7805 does not have this capability, so Analogix provided us this sw
timeout to resolve the problem.

In the case of the timeout, the driver will restart HDCP authentication
in couple seconds. Before this change it would be stuck waiting forever.

Change-Id: If30b08ce3f2f0b674203fc9572e0b756f2a90272
Signed-off-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-on: http://gerrit.mot.com/870006
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>