msm: msm_dba: backup and restore dsi settings before use
With the added ability to reconfigure DSI, we broke the ability to use
a standard DBA function that did not require reconfiguring DSI, because
any new settings from an advanced DBA would still be set. An example of
this is changing the mode from VIDEO to COMMAND. Standard DBA requires
VIDEO mode but the original DBA implementation never intended for that
to change after DTS parsing.

The solution is to backup and restore the DSI driver state before
connect. This will ensure that every time we connect a DBA interface,
the initial state of the DSI is the same.

Change-Id: I650518ef997bdfcc80a30a8bebdd60fbd1e16297
Signed-off-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-on: http://gerrit.mot.com/871522
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Joseph Swantek <jswantek@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>