usb: dwc3: Fix Erratic Error handling.
When the controller generates an erratic error in some scenarios
(such as when a wall charger is incorrectly identified as a SDP),
it generates a cascading chain of Erratic Error Events. The
Error Events cause the controller to be restarted indefinitely and
also prevent low power (since the USB PHY doesn't transition to L2).
We also see the controller left on and fail to transition to lpm,
even after the charger is removed. Fix the above issues by reworking
the error handling to bail out on restarts after 5 attempts. Also
treat the failure to transition the PHY to L2 as a non-fatal error.

Change-Id: Ie648f2494f82ebbeebfb2c70b7b6644424a1d6e3
Signed-off-by: Ravi Vembu <raviv@motorola.com>
Reviewed-on: https://gerrit.mot.com/908775
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>