spi_qsd: Keep BAM ioremap around on removal
Change the ioremap of the BAM physical address to only be done when the
specific instance of the SPI driver will be registering the BAM device
itself. We also purposely do _not_ iounmap the address space as we never
deregister the BAM device and there could be other users.

Commit (da16507 (CR) spi: spi_qsd: Don't deregister BAM)
already removed the sps_deregister_bam_device, however, the devm_ioremap
mapped IO space would be free'd as well when this driver is removed.
Thus, when the next BAM user tries to access the IO address space, the
virtual address would have already be unmapped.

The whole SPS registration here is inherently broken. Each SPI driver
using BAM will try to find an existing BAM device. If there is none, it
becomes responsible for creating it. However, there is no reference
counting of users, so whoever creates the devices is from then on
responsible for keeping it registered. The problem arises when the SPI
driver itself needs to be removed, it normally would free its IO mapped
addresses, of which the BAM device is now using.

This change builds on top of the previous change to keep things working
without re-designing the SPS/BAM driver which is used everywhere.

Change-Id: I88365ff7ec28edfffeea9e07f75286d12e55c0ad
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: https://gerrit.mot.com/919645
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>