msm: camera: sensor: Free power settings on error
Free the power settings and power down settings in the failure path of
the sensor driver probe. If there is a failure probing the device - such
as the device not yet ready - the exit path can leak the allocated
memory. Such a failure is not fatal and is retried in user space.

Change-Id: I9ed14b035fb5bc65b3b5b4320117b6fdb720032e
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: https://gerrit.mot.com/921751
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>