mdss: mot-dba: Reset gpio_sel_dsi when detaching MyDP MOD.
Currently gpio_sel_dsi will be cleared when DSI MOD is attached, and
gpio_sel_dsi will be set for MyDP MOD

Since QCOM has supported to report the HDCP_Capibility which is based
of the gpio_sel_dsi setting. If gpio_sel_dsi is cleared then it will set
HDCP_capability to 0xff, and "01" for gpio_sel_dsi is set.

When the MyDP MOD is removed, gpio_sel_dsi is not cleared and QCOM
will still set HDCP_Capabilioty to "01", which is wrong. This change
will clear the gpio_sel_dsi when myDP is detached.

Change-Id: I67d899a80097a740cd1233af71f13e3f1e0a5d05
Signed-off-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-on: https://gerrit.mot.com/1054975
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Submit-Approved: Jira Key
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>