smb-lib: Clear the USB ICL vote if OTG enable fails
When OTG enable, it will vote USB ICL to be 0mA, and clear USB ICL vote
when OTG disable.
While the USB ICL vote is not clear when OTG enable fails, and it does
not clear it either in OTG disable as OTG is not in enabled state.
In that case, the ICL will be settled to 0mA finally.

So the USB ICL voter from OTG should be cleared if OTG enable fails so
as to avoid this situation.

This is a propagation of (CR) to kernel 3.18.

Change-Id: I54f738c18bac3d4e2c94c65eed55afc335791501
Signed-off-by: Tiago Ganselli <ganselli@motorola.com>
Reviewed-on: https://gerrit.mot.com/1209054
Reviewed-on: https://gerrit.mot.com/1236631
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Konstantin Makariev <kmakariev@motorola.com>
Submit-Approved: Jira Key
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>